+++ packages/testsweets/lib/src/ui/widget_capture/widget_capture_widgets/intro_controllers.dart
@@ -4,6 +4,7 @@ import 'package:testsweets/src/ui/shared/app_colors.dart';
 import 'package:testsweets/src/ui/shared/cta_button.dart';
 
 import '../widget_capture_viewmodel.dart';
+import 'black_wrapper_container.dart';
 
 class IntroControllers extends ViewModelWidget<WidgetCaptureViewModel> {
   const IntroControllers({
@@ -12,23 +13,35 @@ class IntroControllers extends ViewModelWidget<WidgetCaptureViewModel> {
 
   @override
   Widget build(BuildContext context, WidgetCaptureViewModel model) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.end,
-      children: [
-        CtaButton(
-          title: 'Inspect View',
-          fillColor: kcPrimaryFuchsia,
-          onTap: model.toggleInspectLayout,
+    return AnimatedAlign(
+      duration: const Duration(milliseconds: 500),
+      curve: Curves.easeOutCubic,
+      alignment: model.widgetNameInputPositionIsDown
+          ? Alignment.bottomCenter
+          : Alignment.topCenter,
+      widthFactor: 1,
+      child: BlackWrapperContainer(
+        switchPositionTap: model.switchWidgetNameInputPosition,
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          crossAxisAlignment: CrossAxisAlignment.end,
+          children: [
+            CtaButton(
+              title: 'Inspect View',
+              fillColor: kcPrimaryFuchsia,
+              onTap: model.toggleInspectLayout,
+            ),
+            const SizedBox(
+              height: 16,
+            ),
+            CtaButton(
+              title: 'Start Capture',
+              fillColor: kcPrimaryPurple,
+              onTap: model.toggleCaptureView,
+            ),
+          ],
         ),
-        const SizedBox(
-          height: 16,
-        ),
-        CtaButton(
-          title: 'Start Capture',
-          fillColor: kcPrimaryPurple,
-          onTap: model.toggleCaptureView,
-        ),
-      ],
+      ),
     );
   }
 }