+++ packages/testsweets/lib/src/ui/widget_capture/widget_capture_widgets/widget_description_dialog.dart
@@ -1,7 +1,11 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:stacked/stacked.dart';
 import 'package:testsweets/src/ui/shared/shared_styles.dart';
 import 'package:testsweets/src/ui/widget_capture/widget_capture_viewmodel.dart';
+
+import 'black_wrapper_container.dart';
+import 'multi_style_text.dart';
 
 class WidgetDescriptionDialog extends ViewModelWidget<WidgetCaptureViewModel> {
   const WidgetDescriptionDialog({
@@ -12,44 +16,45 @@ class WidgetDescriptionDialog extends ViewModelWidget<WidgetCaptureViewModel> {
   Widget build(BuildContext context, WidgetCaptureViewModel model) {
     final description = model.activeWidgetDescription;
 
-    return Container(
-      height: 250,
-      width: 390,
-      margin: EdgeInsets.symmetric(horizontal: 12),
-      padding: const EdgeInsets.symmetric(horizontal: 15),
+    return BlackWrapperContainer(
+      closeWidget: model.closeWidgetDescription,
+      spaceBetweenTopControllersAndChild: 0,
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
-        mainAxisAlignment: MainAxisAlignment.spaceAround,
         children: [
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Text('Widget Description', style: boldStyle),
-              IconButton(
-                icon: Icon(
-                  Icons.cancel,
-                  color: Colors.grey,
-                  size: 32,
-                ),
-                onPressed: model.closeWidgetDescription,
-              ),
-            ],
+          Text('Widget Description', style: boldStyle),
+          SizedBox(
+            height: 8.h,
           ),
-          if (description != null) ...[
-            Text('View Name: ${description.viewName}', style: lightStyle),
-            Text('Name: ${description.name}', style: lightStyle),
-            Text(
-                'Widget Type: ${description.widgetType.toString().substring(11).capitalizeFirstofEach}',
-                style: lightStyle),
-            Text(
-                'Position: (x:${description.position.x.toStringAsFixed(1)}, y:${description.position.y.toStringAsFixed(1)})',
-                style: lightStyle),
-          ]
+          if (description != null)
+            ...[
+              MultiStyleText(
+                title: 'View Name: ',
+                body: description.viewName,
+              ),
+              MultiStyleText(
+                title: 'Name: ',
+                body: description.name,
+              ),
+              MultiStyleText(
+                title: 'Widget Type: ',
+                body: description.widgetType
+                    .toString()
+                    .substring(11)
+                    .capitalizeFirstofEach,
+              ),
+              MultiStyleText(
+                title: 'Position: ',
+                body:
+                    '( x: ${description.position.x.toStringAsFixed(1)}, y: ${description.position.y.toStringAsFixed(1)} )',
+              ),
+            ].expand((element) => [
+                  element,
+                  SizedBox(
+                    height: 3.h,
+                  )
+                ])
         ],
-      ),
-      decoration: BoxDecoration(
-        color: Color(0xFF181920),
-        borderRadius: BorderRadius.circular(10),
       ),
     );
   }