+++ packages/testsweets/lib/src/ui/widget_capture/widget_capture_widgets/widget_types_container.dart
@@ -43,6 +43,16 @@ class WidgetsTypesContainer extends ViewModelWidget<WidgetCaptureViewModel> {
           SizedBox(
             height: 24.w,
           ),
+          _WidgetTypeButton(
+            onTap: () => model.addNewWidget(WidgetType.scrollable,
+                widgetPosition: WidgetPosition(
+                    x: ScreenUtil().screenWidth / 2,
+                    y: ScreenUtil().screenHeight / 2)),
+            title: 'Scrollable',
+          ),
+          SizedBox(
+            height: 24.w,
+          ),
           Divider(
             color: kcBackground,
             thickness: 5.w,